Modify bootfiles recipe for 64 built builds
Detect if building a 32 or 64 bit image and only modify the kernel entry in the config.txt file if its 32 bit.

64 bit builds must be copied as kernel8.img and in uncompressed form. 
Modified creation scripts for 64 bit images

rpi-config logging 


Amended rpi-config recipe

Conditional adjustment of the kernel image name is not working still
Modified rpi-config_git.bbappend

The output image 'name' in config.txt should only be set to the image type for 32 bit kernels.  For 64 bit kernels it's left blank as the kernel name has to be (currently) kernel8.img

This commit matches a similar change in the card creation scripts